+++ include/behaviortree_cpp/controls/fallback_node.h
@@ -40,7 +40,7 @@ class FallbackNode : public ControlNode
     virtual void halt() override;
 
   private:
-    unsigned int current_child_idx_;
+    size_t current_child_idx_;
 
     virtual BT::NodeStatus tick() override;
 };
